+++ src/main/java/org/folio/mr/service/CirculationRequestService.java
@@ -4,5 +4,7 @@ import org.folio.mr.domain.dto.Request;
 import org.folio.mr.domain.entity.MediatedRequestEntity;
 
 public interface CirculationRequestService {
+  Request get(String id);
   Request create(MediatedRequestEntity mediatedRequest);
+  Request update(Request request);
 }